src/server/requirements.txt
nltk
pandas
flask
flask_cors
google.generativeai
pytrends
statsmodels
requests
urllib3